+++ src/main/java/diplom/work/roomsimulatorservice/dto/LstmModelDTO/LstmModelDtoRequest.java
@@ -3,7 +3,7 @@ package diplom.work.roomsimulatorservice.dto.LstmModelDTO;
 import java.util.List;
 
 public record LstmModelDtoRequest (
-        Integer modelId,
+        Long modelId,
         List<LstmSensorDTO> sensorDataDTOList
 ) {
 }
